+count right sum

+++ karnauhova.alexandra/S1/count_sum.hpp
@@ -6,8 +6,8 @@
 
 namespace karnauhova
 {
-  int element_lists(std::forward_list<int> lists, size_t index);
-  void count_lists(std::forward_list<std::pair<std::string, std::forward_list<int>>> l, std::ostream& out);
+  int element_lists(std::forward_list<unsigned long long> lists, size_t index);
+  void count_lists(std::forward_list<std::pair<std::string, std::forward_list<unsigned long long>>> l, std::ostream& out);
 }
 
 #endif
