+++ karnauhova.alexandra/S1/input.hpp
@@ -7,7 +7,7 @@
 
 namespace karnauhova
 {
-  std::pair<std::string, std::forward_list<int>> input_line(std::istream& in, std::string name);
+  std::pair<std::string, std::forward_list<unsigned long long>> input_line(std::istream& in, std::string name);
 }
 
 #endif
