+++ karnauhova.alexandra/S1/output.cpp
@@ -1,6 +1,6 @@
 #include "output.hpp"
 
-void karnauhova::output_names(std::forward_list<std::pair<std::string, std::forward_list<int>>> l, std::ostream& out)
+void karnauhova::output_names(std::forward_list<std::pair<std::string, std::forward_list<unsigned long long>>> l, std::ostream& out)
 {
   auto it = l.begin();
   if (it == l.end())
@@ -16,7 +16,7 @@ void karnauhova::output_names(std::forward_list<std::pair<std::string, std::forw
   out << "\n";
 }
 
-void karnauhova::output_element_lists(std::forward_list<int> lists, size_t index, std::ostream& out, bool& first)
+void karnauhova::output_element_lists(std::forward_list<unsigned long long> lists, size_t index, std::ostream& out, bool& first)
 {
   size_t count = 1;
   auto it = lists.begin();
@@ -37,7 +37,7 @@ void karnauhova::output_element_lists(std::forward_list<int> lists, size_t index
   out << *it;
 }
 
-void karnauhova::output_lists(std::forward_list<std::pair<std::string, std::forward_list<int>>> l, std::ostream& out)
+void karnauhova::output_lists(std::forward_list<std::pair<std::string, std::forward_list<unsigned long long>>> l, std::ostream& out)
 {
   int max_length = 0;
   for (const auto& it : l)
