+++ karnauhova.alexandra/S1/output.hpp
@@ -6,9 +6,9 @@
 
 namespace karnauhova
 {
-  void output_names(std::forward_list<std::pair<std::string, std::forward_list<int>>> l, std::ostream& out);
-  void output_element_lists(std::forward_list<int> lists, size_t index, std::ostream& out, bool& first);
-  void output_lists(std::forward_list<std::pair<std::string, std::forward_list<int>>> l, std::ostream& out);
+  void output_names(std::forward_list<std::pair<std::string, std::forward_list<unsigned long long>>> l, std::ostream& out);
+  void output_element_lists(std::forward_list<unsigned long long> lists, size_t index, std::ostream& out, bool& first);
+  void output_lists(std::forward_list<std::pair<std::string, std::forward_list<unsigned long long>>> l, std::ostream& out);
 }
 
 #endif
